fix main.js
